Mass Code Rewrite (More info in Description)
Our initial code was a ridiculously cluttered mess.
This became worse as we add more features to the game.
So I thought, there must be something missing... and it was our
lack of planning!
I searched the web for a solution on how to correctly structure our code
without writing the code itself, and this is what I found: UML Class Diagram.
I have studied and applied it for the entire night, and wrote a new code
from scratch but this time, following the UML Class Diagram that I made in Whimsical. 
Surprisingly, it worked so well! that I actually had fun testing out the Game!

TL;DR:
I have rewritten the following classes so far:
-Player
-PlayerManager
-Item
-ItemManager
-ItemSpawner
-Various Interfaces